Reformat README.md for markdown syntax
The README file was written as plain-text, but the extension indicates
that it is to be considered as a Markdown-formatted file. This wouldn't
really be more than a nitpick, if it weren't for the fact that GitHub
uses the README as kind-of a banner on the project page, and the current
syntax renders the links invisible.

As much as anything, this is to learn how to get stuff done with git :)