Add two new loudness fields (integrated_lufs and truepeak_lufs) to VideoFile.
This allow us to measure the loudness for each file and pass this information
on to the playout to adjust the volume of the video during playout to ensure
we send with LUFS -23 according to the Norwegian TV sound level standard.

The fields are optional, because not all VideoFile entries are videos,
some are images and subtitles.

Fixes #95