Remove extraneous text
With the development focus now being on getting the "old" stack running and maintainable, and most of the running code for that still being here, I propose we cancel the notion of migrating away from the monorepo.

I also removed some obsolete text. The README is not the appropriate place for project management anyway.